helperrvv.h refactor
For functions which conflict between SP and DP configurations, use
distinct type and function names, and then use the preprocessor to
decide which ones will be available under the standard names.  This
works better after preprocessing for inline headers, and allows internal
references to the SP or DP functions when needed by other functions.

Prefer vcreate intrinsics because they'll define the whole data type
without any temporary uninitialised state.  Except older compilers don't
support this so we need a fallback.

When padding RVV registers to a wider format in debug builds, use an
extra move operation so that qemu can mark the undefined part of the
register.